% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_plot.R
\name{compare_nodes}
\alias{compare_nodes}
\title{Compares the output of several nodes}
\usage{
compare_nodes(node_list, type = "density")
}
\description{
Different from plot_outputs() because it does not check dependencies. Instead,
nodes are passed explicitly. Therefore, it can be used to compare between different
models.
}
