% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response.R
\name{DoseResponse_Exponential}
\alias{DoseResponse_Exponential}
\title{R6 class describing the exponential dose-response model}
\description{
R6 class describing the exponential dose-response model

R6 class describing the exponential dose-response model
}
\details{
A risk module describing the exponential dose-response model. It has 2 inputs: r, dose
}
\section{Super classes}{
\code{\link[biorisk:RiskModule]{biorisk::RiskModule}} -> \code{biorisk::NA} -> \code{DoseResponse_Exponential}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DoseResponse_Exponential-new}{\code{DoseResponse_Exponential$new()}}
\item \href{#method-DoseResponse_Exponential-point_estimate}{\code{DoseResponse_Exponential$point_estimate()}}
\item \href{#method-DoseResponse_Exponential-clone}{\code{DoseResponse_Exponential$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="boxplot"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-boxplot'><code>biorisk::RiskModule$boxplot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="check_input_types"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-check_input_types'><code>biorisk::RiskModule$check_input_types()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="cummulative_plot"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-cummulative_plot'><code>biorisk::RiskModule$cummulative_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="cummulative_plot_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-cummulative_plot_2D'><code>biorisk::RiskModule$cummulative_plot_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="density_plot"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-density_plot'><code>biorisk::RiskModule$density_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="density_plot_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-density_plot_2D'><code>biorisk::RiskModule$density_plot_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="get_output"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-get_output'><code>biorisk::RiskModule$get_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="get_output_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-get_output_2D'><code>biorisk::RiskModule$get_output_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="get_output_type"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-get_output_type'><code>biorisk::RiskModule$get_output_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="histogram"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-histogram'><code>biorisk::RiskModule$histogram()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="map_input"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-map_input'><code>biorisk::RiskModule$map_input()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="quantiles"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-quantiles'><code>biorisk::RiskModule$quantiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="quantiles_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-quantiles_2D'><code>biorisk::RiskModule$quantiles_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="save_as_distribution"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-save_as_distribution'><code>biorisk::RiskModule$save_as_distribution()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="save_as_vector"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-save_as_vector'><code>biorisk::RiskModule$save_as_vector()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="simulate"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-simulate'><code>biorisk::RiskModule$simulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="simulate_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-simulate_2D'><code>biorisk::RiskModule$simulate_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="simulate_level"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-simulate_level'><code>biorisk::RiskModule$simulate_level()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DoseResponse_Exponential-new"></a>}}
\if{latex}{\out{\hypertarget{method-DoseResponse_Exponential-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DoseResponse_Exponential$new(name, units = NA, output_unit = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DoseResponse_Exponential-point_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-DoseResponse_Exponential-point_estimate}{}}}
\subsection{Method \code{point_estimate()}}{
Returns the expected value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DoseResponse_Exponential$point_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DoseResponse_Exponential-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DoseResponse_Exponential-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DoseResponse_Exponential$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
