% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_inactivation.R
\name{DynamicBigelow_2phase}
\alias{DynamicBigelow_2phase}
\title{R6 class describing log-linear inactivation under dynamic conditions for a ramp}
\description{
R6 class describing log-linear inactivation under dynamic conditions for a ramp

R6 class describing log-linear inactivation under dynamic conditions for a ramp
}
\details{
A risk element describing log-linear inactivation under dynamic conditions considering
a heating (or cooling) profile. It has 9 inputs: t1, t2, D_R, z, temp_ref, logN0, temp0, temp1, temp2,
}
\section{Super classes}{
\code{biorisk::RiskElement} -> \code{biorisk::NA} -> \code{DynamicBigelow_2phase}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DynamicBigelow_2phase-new}{\code{DynamicBigelow_2phase$new()}}
\item \href{#method-DynamicBigelow_2phase-point_estimate}{\code{DynamicBigelow_2phase$point_estimate()}}
\item \href{#method-DynamicBigelow_2phase-clone}{\code{DynamicBigelow_2phase$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="boxplot"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-boxplot'><code>biorisk::RiskElement$boxplot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="check_input_types"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-check_input_types'><code>biorisk::RiskElement$check_input_types()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="cummulative_plot"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-cummulative_plot'><code>biorisk::RiskElement$cummulative_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="cummulative_plot_2D"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-cummulative_plot_2D'><code>biorisk::RiskElement$cummulative_plot_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="density_plot"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-density_plot'><code>biorisk::RiskElement$density_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="density_plot_2D"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-density_plot_2D'><code>biorisk::RiskElement$density_plot_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="get_output"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-get_output'><code>biorisk::RiskElement$get_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="get_output_2D"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-get_output_2D'><code>biorisk::RiskElement$get_output_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="get_output_type"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-get_output_type'><code>biorisk::RiskElement$get_output_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="histogram"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-histogram'><code>biorisk::RiskElement$histogram()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="map_input"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-map_input'><code>biorisk::RiskElement$map_input()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="quantiles"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-quantiles'><code>biorisk::RiskElement$quantiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="quantiles_2D"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-quantiles_2D'><code>biorisk::RiskElement$quantiles_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="save_as_distribution"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-save_as_distribution'><code>biorisk::RiskElement$save_as_distribution()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="save_as_vector"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-save_as_vector'><code>biorisk::RiskElement$save_as_vector()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="simulate"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-simulate'><code>biorisk::RiskElement$simulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="simulate_2D"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-simulate_2D'><code>biorisk::RiskElement$simulate_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="simulate_level"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-simulate_level'><code>biorisk::RiskElement$simulate_level()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynamicBigelow_2phase-new"></a>}}
\if{latex}{\out{\hypertarget{method-DynamicBigelow_2phase-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicBigelow_2phase$new(name, units = NA, output_unit = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynamicBigelow_2phase-point_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-DynamicBigelow_2phase-point_estimate}{}}}
\subsection{Method \code{point_estimate()}}{
Returns the expected value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicBigelow_2phase$point_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynamicBigelow_2phase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DynamicBigelow_2phase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynamicBigelow_2phase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
