% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{ExponentialGrowth}
\alias{ExponentialGrowth}
\title{Growth module based on exponential growth}
\description{
Growth module based on exponential growth

Growth module based on exponential growth
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExponentialGrowth$new()}}
\item \href{#method-map_input}{\code{ExponentialGrowth$map_input()}}
\item \href{#method-simulate}{\code{ExponentialGrowth$simulate()}}
\item \href{#method-get_output}{\code{ExponentialGrowth$get_output()}}
\item \href{#method-clone}{\code{ExponentialGrowth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialGrowth$new(name, units = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-map_input"></a>}}
\if{latex}{\out{\hypertarget{method-map_input}{}}}
\subsection{Method \code{map_input()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialGrowth$map_input(input, module, check_units = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-simulate}{}}}
\subsection{Method \code{simulate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialGrowth$simulate(niter)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_output"></a>}}
\if{latex}{\out{\hypertarget{method-get_output}{}}}
\subsection{Method \code{get_output()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialGrowth$get_output()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExponentialGrowth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
