% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algebra.R
\name{ModuleExp}
\alias{ModuleExp}
\title{Exponential of the output of a module}
\description{
Exponential of the output of a module

Exponential of the output of a module
}
\section{Super class}{
\code{\link[biorisk:RiskModule]{biorisk::RiskModule}} -> \code{ModuleExp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModuleExp-new}{\code{ModuleExp$new()}}
\item \href{#method-ModuleExp-get_output_type}{\code{ModuleExp$get_output_type()}}
\item \href{#method-ModuleExp-point_estimate}{\code{ModuleExp$point_estimate()}}
\item \href{#method-ModuleExp-clone}{\code{ModuleExp$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="boxplot"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-boxplot'><code>biorisk::RiskModule$boxplot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="check_input_types"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-check_input_types'><code>biorisk::RiskModule$check_input_types()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="cummulative_plot"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-cummulative_plot'><code>biorisk::RiskModule$cummulative_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="cummulative_plot_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-cummulative_plot_2D'><code>biorisk::RiskModule$cummulative_plot_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="density_plot"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-density_plot'><code>biorisk::RiskModule$density_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="density_plot_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-density_plot_2D'><code>biorisk::RiskModule$density_plot_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="get_output"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-get_output'><code>biorisk::RiskModule$get_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="get_output_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-get_output_2D'><code>biorisk::RiskModule$get_output_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="histogram"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-histogram'><code>biorisk::RiskModule$histogram()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="map_input"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-map_input'><code>biorisk::RiskModule$map_input()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="quantiles"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-quantiles'><code>biorisk::RiskModule$quantiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="quantiles_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-quantiles_2D'><code>biorisk::RiskModule$quantiles_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="save_as_distribution"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-save_as_distribution'><code>biorisk::RiskModule$save_as_distribution()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="save_as_vector"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-save_as_vector'><code>biorisk::RiskModule$save_as_vector()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="simulate"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-simulate'><code>biorisk::RiskModule$simulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="simulate_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-simulate_2D'><code>biorisk::RiskModule$simulate_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="simulate_level"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-simulate_level'><code>biorisk::RiskModule$simulate_level()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleExp-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleExp-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleExp$new(name, units = NA, output_unit = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleExp-get_output_type"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleExp-get_output_type}{}}}
\subsection{Method \code{get_output_type()}}{
Get the data type of the output.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleExp$get_output_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleExp-point_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleExp-point_estimate}{}}}
\subsection{Method \code{point_estimate()}}{
Returns the expected value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleExp$point_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModuleExp-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModuleExp-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModuleExp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
