% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_inactivation.R
\name{predict_Bigelow_biphasic}
\alias{predict_Bigelow_biphasic}
\title{Dynamic microbial inactivation for the Bigelow model with a biphasic profile}
\usage{
predict_Bigelow_biphasic(t1, t2, D_R, z, temp_ref, logN0, temp0, temp1, temp2)
}
\description{
Dynamic microbial inactivation for the Bigelow model with a biphasic profile
}
