% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_inactivation.R
\name{predict_Bigelow_ramp}
\alias{predict_Bigelow_ramp}
\title{Dynamic microbial inactivation for the Bigelow model with a ramp profile}
\usage{
predict_Bigelow_ramp(max_time, D_R, z, temp_ref, logN0, temp_start, temp_end)
}
\description{
Dynamic microbial inactivation for the Bigelow model with a ramp profile
}
