% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_inactivation.R
\name{predict_Bigelow_triphasic}
\alias{predict_Bigelow_triphasic}
\title{Dynamic microbial inactivation for the Bigelow model with a triphasic profile}
\usage{
predict_Bigelow_triphasic(
  t1,
  t2,
  t3,
  D_R,
  z,
  temp_ref,
  logN0,
  temp0,
  temp1,
  temp2,
  temp3
)
}
\description{
Dynamic microbial inactivation for the Bigelow model with a triphasic profile
}
