% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_summaries.R
\name{quantile_table}
\alias{quantile_table}
\title{Returns a table of the model quantiles}
\usage{
quantile_table(node, probs = seq(0, 1, 0.25), chosen = NULL)
}
\description{
Returns a table of the model quantiles
}
