% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{regression_tree}
\alias{regression_tree}
\title{Regression tree of the output}
\usage{
regression_tree(node, chosen = NULL, make_plot = FALSE)
}
\arguments{
\item{node}{The output node.}

\item{chosen}{A character vector of node names to include in the tree. By default, \code{NULL} (every node).}

\item{make_plot}{Whether to plot the tree. \code{FALSE} by default.}
}
\description{
Useful for sensitivity analysis
}
