% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_convergence.R
\name{resample_vector_median}
\alias{resample_vector_median}
\title{Resampling a vector with replacement several times}
\usage{
resample_vector_median(x, size, samples)
}
\description{
Resampling a vector with replacement several times
}
