% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_plot.R
\name{retrieve_outputs}
\alias{retrieve_outputs}
\title{Gets the output of all the nodes of a model}
\usage{
retrieve_outputs(node)
}
\arguments{
\item{node}{base node}
}
\description{
Gets the output of all the nodes of a model
}
