% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_summaries.R
\name{retrieve_table}
\alias{retrieve_table}
\title{Builds a table with quantiles of the model outputs}
\usage{
retrieve_table(node, probs = seq(0, 1, 0.25))
}
\description{
Builds a table with quantiles of the model outputs
}
