% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_module.R
\name{RiskElement}
\alias{RiskElement}
\title{Parent class for elements}
\description{
Parent class for elements

Parent class for elements
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character giving a name to this constant}

\item{\code{inputs}}{A tibble with the inputs of the element and its units}

\item{\code{depends_on}}{List describing the dependencies for each input}

\item{\code{depends_index}}{List with the indexes for each independency}

\item{\code{depended_by}}{A list of elements that use this instance as input}

\item{\code{input_types}}{A list with the preferred data type for each input (discrete/continuous/any)}

\item{\code{output}}{Name of the output variable}

\item{\code{output_unit}}{Unit of the output variable}

\item{\code{output_type}}{Type of output variable (discrete/continuous)}

\item{\code{simulations}}{A tibble with the results of the simulations}

\item{\code{type}}{A character describing the type of the element}

\item{\code{level}}{Description of the level for X-D Monte Carlo}

\item{\code{simulations_multi}}{For multidimensional MC}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RiskElement-new}{\code{RiskElement$new()}}
\item \href{#method-RiskElement-map_input}{\code{RiskElement$map_input()}}
\item \href{#method-RiskElement-simulate}{\code{RiskElement$simulate()}}
\item \href{#method-RiskElement-simulate_2D}{\code{RiskElement$simulate_2D()}}
\item \href{#method-RiskElement-simulate_level}{\code{RiskElement$simulate_level()}}
\item \href{#method-RiskElement-point_estimate}{\code{RiskElement$point_estimate()}}
\item \href{#method-RiskElement-get_output}{\code{RiskElement$get_output()}}
\item \href{#method-RiskElement-get_output_2D}{\code{RiskElement$get_output_2D()}}
\item \href{#method-RiskElement-save_as_vector}{\code{RiskElement$save_as_vector()}}
\item \href{#method-RiskElement-save_as_distribution}{\code{RiskElement$save_as_distribution()}}
\item \href{#method-RiskElement-density_plot_2D}{\code{RiskElement$density_plot_2D()}}
\item \href{#method-RiskElement-cummulative_plot_2D}{\code{RiskElement$cummulative_plot_2D()}}
\item \href{#method-RiskElement-cummulative_plot}{\code{RiskElement$cummulative_plot()}}
\item \href{#method-RiskElement-density_plot}{\code{RiskElement$density_plot()}}
\item \href{#method-RiskElement-histogram}{\code{RiskElement$histogram()}}
\item \href{#method-RiskElement-boxplot}{\code{RiskElement$boxplot()}}
\item \href{#method-RiskElement-quantiles}{\code{RiskElement$quantiles()}}
\item \href{#method-RiskElement-quantiles_2D}{\code{RiskElement$quantiles_2D()}}
\item \href{#method-RiskElement-get_output_type}{\code{RiskElement$get_output_type()}}
\item \href{#method-RiskElement-check_input_types}{\code{RiskElement$check_input_types()}}
\item \href{#method-RiskElement-clone}{\code{RiskElement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-new"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-new}{}}}
\subsection{Method \code{new()}}{
Create a new element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$new(
  name,
  input_names = NA,
  units = NA,
  element_type = "",
  input_types = list(),
  output_var = "",
  output_unit = "",
  output_type = "",
  level = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name}

\item{\code{input_names}}{A character vector with the names of the inputs}

\item{\code{units}}{A character vector of units for each input}

\item{\code{element_type}}{A character with the type of element}

\item{\code{output_var}}{A character with the name of the output variable}

\item{\code{output_unit}}{A character with the unit of the output}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-map_input"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-map_input}{}}}
\subsection{Method \code{map_input()}}{
Map an input to a element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$map_input(input, element, check_units = FALSE, index = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{A character identifying the input variable}

\item{\code{element}}{An instance of a element to use as input}

\item{\code{check_units}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self (invisible)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-simulate}{}}}
\subsection{Method \code{simulate()}}{
Make simulation. Returns a vector of length niter with self.value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$simulate(niter, seed = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{niter}}{Number of iterations (length of the vector).}

\item{\code{check_units}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector with the output variable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-simulate_2D"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-simulate_2D}{}}}
\subsection{Method \code{simulate_2D()}}{
2D Monte Carlo simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$simulate_2D(niter0, niter1, seed = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{niter0}}{number of iterations on the lower level}

\item{\code{niter1}}{number of iterations on the upper level}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-simulate_level"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-simulate_level}{}}}
\subsection{Method \code{simulate_level()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$simulate_level(niter0, iter1 = 1, level = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-point_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-point_estimate}{}}}
\subsection{Method \code{point_estimate()}}{
Gets a discrete (fast and simple) prediction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$point_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-get_output"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-get_output}{}}}
\subsection{Method \code{get_output()}}{
Get the output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$get_output(iter1 = NULL, index = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{niter}}{Number of iterations (length of the vector).}

\item{\code{check_units}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector with the output variable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-get_output_2D"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-get_output_2D}{}}}
\subsection{Method \code{get_output_2D()}}{
Get the output of a 2D simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$get_output_2D(index = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-save_as_vector"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-save_as_vector}{}}}
\subsection{Method \code{save_as_vector()}}{
Saves the output of the simulation as a vector element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$save_as_vector(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the new element (vctr_from_+self_name by default)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of Vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-save_as_distribution"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-save_as_distribution}{}}}
\subsection{Method \code{save_as_distribution()}}{
Saves the output of the simulation as an empirical distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$save_as_distribution(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the new element (distr_from_+self_name by default)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of EmpiricalDistr
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-density_plot_2D"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-density_plot_2D}{}}}
\subsection{Method \code{density_plot_2D()}}{
Makes density plot of a 2D Monte Carlo simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$density_plot_2D()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-cummulative_plot_2D"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-cummulative_plot_2D}{}}}
\subsection{Method \code{cummulative_plot_2D()}}{
Plots the empirical density function for 2D simulations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$cummulative_plot_2D()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-cummulative_plot"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-cummulative_plot}{}}}
\subsection{Method \code{cummulative_plot()}}{
Plots the empirical density function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$cummulative_plot(add_discrete = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-density_plot"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-density_plot}{}}}
\subsection{Method \code{density_plot()}}{
Makes a density plot of the model output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$density_plot(add_discrete = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-histogram"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-histogram}{}}}
\subsection{Method \code{histogram()}}{
Makes a histogram of the model output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$histogram(add_discrete = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-boxplot"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-boxplot}{}}}
\subsection{Method \code{boxplot()}}{
Makes a boxplot of the model output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$boxplot(add_discrete = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-quantiles"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-quantiles}{}}}
\subsection{Method \code{quantiles()}}{
Quantiles of the output variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$quantiles(probs = c(0.01, 0.1, 0.5, 0.9, 0.99))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-quantiles_2D"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-quantiles_2D}{}}}
\subsection{Method \code{quantiles_2D()}}{
Quantiles of a 2D Monte Carlo simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$quantiles_2D(probs = c(0.01, 0.1, 0.5, 0.9, 0.99))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-get_output_type"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-get_output_type}{}}}
\subsection{Method \code{get_output_type()}}{
Get the data type of the output.
This is a default implementation (just return the output type). For other elements,
(e.g., sum) the type depends on the type of the inputs. So, they have their
own implementation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$get_output_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-check_input_types"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-check_input_types}{}}}
\subsection{Method \code{check_input_types()}}{
Checks that the type of the inputs is consistent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$check_input_types(recursive = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskElement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RiskElement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
