% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent_module.R
\name{RiskModule}
\alias{RiskModule}
\title{Parent class for modules}
\description{
Parent class for modules

Parent class for modules
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character giving a name to this constant}

\item{\code{inputs}}{A tibble with the inputs of the module and its units}

\item{\code{depends_on}}{List describing the dependencies for each input}

\item{\code{depends_index}}{List with the indexes for each independency}

\item{\code{depended_by}}{A list of modules that use this instance as input}

\item{\code{output}}{Name of the output variable}

\item{\code{output_unit}}{Unit of the output variable}

\item{\code{output_type}}{Type of output variable (discrete/continuous)}

\item{\code{simulations}}{A tibble with the results of the simulations}

\item{\code{type}}{A character describing the type of the module}

\item{\code{level}}{Description of the level for X-D Monte Carlo}

\item{\code{simulations_multi}}{For multidimensional MC}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RiskModule-new}{\code{RiskModule$new()}}
\item \href{#method-RiskModule-map_input}{\code{RiskModule$map_input()}}
\item \href{#method-RiskModule-simulate}{\code{RiskModule$simulate()}}
\item \href{#method-RiskModule-simulate_2D}{\code{RiskModule$simulate_2D()}}
\item \href{#method-RiskModule-simulate_level}{\code{RiskModule$simulate_level()}}
\item \href{#method-RiskModule-discrete_prediction}{\code{RiskModule$discrete_prediction()}}
\item \href{#method-RiskModule-get_output}{\code{RiskModule$get_output()}}
\item \href{#method-RiskModule-get_output_2D}{\code{RiskModule$get_output_2D()}}
\item \href{#method-RiskModule-save_as_vector}{\code{RiskModule$save_as_vector()}}
\item \href{#method-RiskModule-save_as_distribution}{\code{RiskModule$save_as_distribution()}}
\item \href{#method-RiskModule-density_plot_2D}{\code{RiskModule$density_plot_2D()}}
\item \href{#method-RiskModule-cummulative_plot_2D}{\code{RiskModule$cummulative_plot_2D()}}
\item \href{#method-RiskModule-cummulative_plot}{\code{RiskModule$cummulative_plot()}}
\item \href{#method-RiskModule-density_plot}{\code{RiskModule$density_plot()}}
\item \href{#method-RiskModule-histogram}{\code{RiskModule$histogram()}}
\item \href{#method-RiskModule-boxplot}{\code{RiskModule$boxplot()}}
\item \href{#method-RiskModule-quantiles}{\code{RiskModule$quantiles()}}
\item \href{#method-RiskModule-quantiles_2D}{\code{RiskModule$quantiles_2D()}}
\item \href{#method-RiskModule-clone}{\code{RiskModule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-new}{}}}
\subsection{Method \code{new()}}{
Create a new module
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$new(
  name,
  input_names = NA,
  units = NA,
  module_type = "",
  output_var = "",
  output_unit = "",
  output_type = "",
  level = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name}

\item{\code{input_names}}{A character vector with the names of the inputs}

\item{\code{units}}{A character vector of units for each input}

\item{\code{module_type}}{A character with the type of module}

\item{\code{output_var}}{A character with the name of the output variable}

\item{\code{output_unit}}{A character with the unit of the output}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the module
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-map_input"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-map_input}{}}}
\subsection{Method \code{map_input()}}{
Map an input to a module
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$map_input(input, module, check_units = FALSE, index = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{A character identifying the input variable}

\item{\code{module}}{An instance of a module to use as input}

\item{\code{check_units}}{Ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self (invisible)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-simulate}{}}}
\subsection{Method \code{simulate()}}{
Make simulation. Returns a vector of length niter with self.value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$simulate(niter, seed = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{niter}}{Number of iterations (length of the vector).}

\item{\code{check_units}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector with the output variable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-simulate_2D"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-simulate_2D}{}}}
\subsection{Method \code{simulate_2D()}}{
2D Monte Carlo simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$simulate_2D(niter0, niter1, seed = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{niter0}}{number of iterations on the lower level}

\item{\code{niter1}}{number of iterations on the upper level}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-simulate_level"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-simulate_level}{}}}
\subsection{Method \code{simulate_level()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$simulate_level(niter0, iter1 = 1, level = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-discrete_prediction"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-discrete_prediction}{}}}
\subsection{Method \code{discrete_prediction()}}{
Gets a discrete (fast and simple) prediction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$discrete_prediction()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-get_output"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-get_output}{}}}
\subsection{Method \code{get_output()}}{
Get the output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$get_output(iter1 = NULL, index = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{niter}}{Number of iterations (length of the vector).}

\item{\code{check_units}}{Ignored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector with the output variable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-get_output_2D"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-get_output_2D}{}}}
\subsection{Method \code{get_output_2D()}}{
Get the output of a 2D simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$get_output_2D(index = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-save_as_vector"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-save_as_vector}{}}}
\subsection{Method \code{save_as_vector()}}{
Saves the output of the simulation as a vector module
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$save_as_vector(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the new module (vctr_from_+self_name by default)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of Vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-save_as_distribution"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-save_as_distribution}{}}}
\subsection{Method \code{save_as_distribution()}}{
Saves the output of the simulation as an empirical distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$save_as_distribution(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the new module (distr_from_+self_name by default)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of EmpiricalDistr
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-density_plot_2D"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-density_plot_2D}{}}}
\subsection{Method \code{density_plot_2D()}}{
Makes density plot of a 2D Monte Carlo simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$density_plot_2D()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-cummulative_plot_2D"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-cummulative_plot_2D}{}}}
\subsection{Method \code{cummulative_plot_2D()}}{
Plots the empirical density function for 2D simulations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$cummulative_plot_2D()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-cummulative_plot"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-cummulative_plot}{}}}
\subsection{Method \code{cummulative_plot()}}{
Plots the empirical density function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$cummulative_plot(add_discrete = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-density_plot"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-density_plot}{}}}
\subsection{Method \code{density_plot()}}{
Makes a density plot of the model output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$density_plot(add_discrete = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-histogram"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-histogram}{}}}
\subsection{Method \code{histogram()}}{
Makes a histogram of the model output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$histogram(add_discrete = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-boxplot"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-boxplot}{}}}
\subsection{Method \code{boxplot()}}{
Makes a boxplot of the model output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$boxplot(add_discrete = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-quantiles"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-quantiles}{}}}
\subsection{Method \code{quantiles()}}{
Quantiles of the output variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$quantiles(probs = c(0.01, 0.1, 0.5, 0.9, 0.99))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-quantiles_2D"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-quantiles_2D}{}}}
\subsection{Method \code{quantiles_2D()}}{
Quantiles of a 2D Monte Carlo simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$quantiles_2D(probs = c(0.01, 0.1, 0.5, 0.9, 0.99))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiskModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RiskModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiskModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
