% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multim_MC.R
\name{simulate_2D}
\alias{simulate_2D}
\title{2D Monte Carlo simulation}
\usage{
simulate_2D(node, niter0, niter1)
}
\arguments{
\item{node}{The base node for the simulations}

\item{niter0}{Number of iterations on the base level}

\item{niter1}{Number of iterations on the upper level}
}
\description{
2D Monte Carlo simulation
}
