% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{Vector}
\alias{Vector}
\title{R6 Class Representing a Constant vector}
\description{
A constant vector that can be used as input by other methods
}
\section{Super class}{
\code{\link[biorisk:RiskModule]{biorisk::RiskModule}} -> \code{Vector}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{The value of the constant}

\item{\code{n}}{Length of the vector}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{value}}{The value of the constant}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Vector-new}{\code{Vector$new()}}
\item \href{#method-Vector-simulate}{\code{Vector$simulate()}}
\item \href{#method-Vector-simulate_level}{\code{Vector$simulate_level()}}
\item \href{#method-Vector-clone}{\code{Vector$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="boxplot"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-boxplot'><code>biorisk::RiskModule$boxplot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="cummulative_plot"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-cummulative_plot'><code>biorisk::RiskModule$cummulative_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="cummulative_plot_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-cummulative_plot_2D'><code>biorisk::RiskModule$cummulative_plot_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="density_plot"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-density_plot'><code>biorisk::RiskModule$density_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="density_plot_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-density_plot_2D'><code>biorisk::RiskModule$density_plot_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="discrete_prediction"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-discrete_prediction'><code>biorisk::RiskModule$discrete_prediction()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="get_output"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-get_output'><code>biorisk::RiskModule$get_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="get_output_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-get_output_2D'><code>biorisk::RiskModule$get_output_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="histogram"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-histogram'><code>biorisk::RiskModule$histogram()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="map_input"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-map_input'><code>biorisk::RiskModule$map_input()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="quantiles"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-quantiles'><code>biorisk::RiskModule$quantiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="quantiles_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-quantiles_2D'><code>biorisk::RiskModule$quantiles_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="save_as_distribution"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-save_as_distribution'><code>biorisk::RiskModule$save_as_distribution()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="save_as_vector"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-save_as_vector'><code>biorisk::RiskModule$save_as_vector()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="simulate_2D"><a href='../../biorisk/html/RiskModule.html#method-RiskModule-simulate_2D'><code>biorisk::RiskModule$simulate_2D()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vector-new"></a>}}
\if{latex}{\out{\hypertarget{method-Vector-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vector$new(name, values, output_unit = NA, level = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vector-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-Vector-simulate}{}}}
\subsection{Method \code{simulate()}}{
Make simulation. Returns the vector saved in the class. If the length
is different from niter, it raises an error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vector$simulate(niter, check_units = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{niter}}{Number of iterations (length of the vector).}

\item{\code{check_units}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vector-simulate_level"></a>}}
\if{latex}{\out{\hypertarget{method-Vector-simulate_level}{}}}
\subsection{Method \code{simulate_level()}}{
Simulation at some level for X-D Monte Carlo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vector$simulate_level(niter0, iter1 = 1, level = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{niter0}}{Number of iterations at the lower level}

\item{\code{iter1}}{Iteration at the upper level}

\item{\code{level}}{Level of the calculation (ignored, it always returns a vector
to facilitate the mutates)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Vector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
