% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{Constant}
\alias{Constant}
\title{R6 Class Representing a Constant}
\description{
A constant that can be used as input by other methods
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character giving a name to this constant}

\item{\code{inputs}}{Ignored (empty tibble)}

\item{\code{depends_on}}{Ignored (empty list)}

\item{\code{depended_by}}{A list of modules that use this instance as input}

\item{\code{output}}{Ignored}

\item{\code{simulations}}{Ignored}

\item{\code{type}}{A character defining it as a constant}

\item{\code{unit}}{The unit of the constant}

\item{\code{value}}{The value of the constant}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Constant$new()}}
\item \href{#method-simulate}{\code{Constant$simulate()}}
\item \href{#method-clone}{\code{Constant$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new constant
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Constant$new(name, value, unit = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name}

\item{\code{value}}{Value of the constant}

\item{\code{unit}}{Unit of the constant (NA by default").}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Constant object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-simulate}{}}}
\subsection{Method \code{simulate()}}{
Make simulation. Returns a vector of length niter with self.value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Constant$simulate(niter, check_units = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{niter}}{Number of iterations (length of the vector).}

\item{\code{check_units}}{Ignore.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Constant$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
