% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inactivation_secondary.R
\name{Peleg_secondary}
\alias{Peleg_secondary}
\title{R6 class describing the secondary model for the Peleg model}
\description{
R6 class describing the secondary model for the Peleg model

R6 class describing the secondary model for the Peleg model
}
\details{
A risk module describing the secondary model of the Peleg model. It has 3 inputs:
k_b, temperature, temp_crit.
}
\section{Super classes}{
\code{\link[biorisk:RiskModule]{biorisk::RiskModule}} -> \code{\link[biorisk:NA]{biorisk::NA}} -> \code{Peleg_secondary}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Peleg_secondary$new()}}
\item \href{#method-discrete_prediction}{\code{Peleg_secondary$discrete_prediction()}}
\item \href{#method-clone}{\code{Peleg_secondary$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="boxplot">}\href{../../biorisk/html/RiskModule.html#method-boxplot}{\code{biorisk::RiskModule$boxplot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="density_plot">}\href{../../biorisk/html/RiskModule.html#method-density_plot}{\code{biorisk::RiskModule$density_plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="get_output">}\href{../../biorisk/html/RiskModule.html#method-get_output}{\code{biorisk::RiskModule$get_output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="histogram">}\href{../../biorisk/html/RiskModule.html#method-histogram}{\code{biorisk::RiskModule$histogram()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="map_input">}\href{../../biorisk/html/RiskModule.html#method-map_input}{\code{biorisk::RiskModule$map_input()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="quantiles">}\href{../../biorisk/html/RiskModule.html#method-quantiles}{\code{biorisk::RiskModule$quantiles()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="save_as_distribution">}\href{../../biorisk/html/RiskModule.html#method-save_as_distribution}{\code{biorisk::RiskModule$save_as_distribution()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="save_as_vector">}\href{../../biorisk/html/RiskModule.html#method-save_as_vector}{\code{biorisk::RiskModule$save_as_vector()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="simulate">}\href{../../biorisk/html/RiskModule.html#method-simulate}{\code{biorisk::RiskModule$simulate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biorisk" data-topic="RiskModule" data-id="simulate_level">}\href{../../biorisk/html/RiskModule.html#method-simulate_level}{\code{biorisk::RiskModule$simulate_level()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Peleg_secondary$new(name, units = NA, output_unit = NA, level = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-discrete_prediction"></a>}}
\if{latex}{\out{\hypertarget{method-discrete_prediction}{}}}
\subsection{Method \code{discrete_prediction()}}{
Returns the expected value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Peleg_secondary$discrete_prediction()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Peleg_secondary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
