% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{Vector}
\alias{Vector}
\title{Vector Class}
\description{
An element that defines a constant vector
}
\section{Super class}{
\code{\link[biorisk:RiskElement]{biorisk::RiskElement}} -> \code{Vector}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{values}}{The values of the vector}

\item{\code{n}}{Length of the vector}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Vector-new}{\code{Vector$new()}}
\item \href{#method-Vector-simulate}{\code{Vector$simulate()}}
\item \href{#method-Vector-simulate_level}{\code{Vector$simulate_level()}}
\item \href{#method-Vector-clone}{\code{Vector$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="boxplot"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-boxplot'><code>biorisk::RiskElement$boxplot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="check_input_types"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-check_input_types'><code>biorisk::RiskElement$check_input_types()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="cummulative_plot"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-cummulative_plot'><code>biorisk::RiskElement$cummulative_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="cummulative_plot_2D"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-cummulative_plot_2D'><code>biorisk::RiskElement$cummulative_plot_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="density_plot"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-density_plot'><code>biorisk::RiskElement$density_plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="density_plot_2D"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-density_plot_2D'><code>biorisk::RiskElement$density_plot_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="get_output"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-get_output'><code>biorisk::RiskElement$get_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="get_output_2D"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-get_output_2D'><code>biorisk::RiskElement$get_output_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="get_output_type"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-get_output_type'><code>biorisk::RiskElement$get_output_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="histogram"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-histogram'><code>biorisk::RiskElement$histogram()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="map_input"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-map_input'><code>biorisk::RiskElement$map_input()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="point_estimate"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-point_estimate'><code>biorisk::RiskElement$point_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="quantiles"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-quantiles'><code>biorisk::RiskElement$quantiles()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="quantiles_2D"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-quantiles_2D'><code>biorisk::RiskElement$quantiles_2D()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="save_as_distribution"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-save_as_distribution'><code>biorisk::RiskElement$save_as_distribution()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="save_as_vector"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-save_as_vector'><code>biorisk::RiskElement$save_as_vector()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biorisk" data-topic="RiskElement" data-id="simulate_2D"><a href='../../biorisk/html/RiskElement.html#method-RiskElement-simulate_2D'><code>biorisk::RiskElement$simulate_2D()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vector-new"></a>}}
\if{latex}{\out{\hypertarget{method-Vector-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vector$new(name, values, output_unit = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A character defining the name for the element}

\item{\code{values}}{A vector of values for the vector}

\item{\code{output_unit}}{A character with the unit of the output}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vector-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-Vector-simulate}{}}}
\subsection{Method \code{simulate()}}{
Make simulation. Returns the vector saved in the class. If the length
is different from niter, it raises an error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vector$simulate(niter, check_units = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{niter}}{Number of iterations (length of the vector).}

\item{\code{check_units}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vector-simulate_level"></a>}}
\if{latex}{\out{\hypertarget{method-Vector-simulate_level}{}}}
\subsection{Method \code{simulate_level()}}{
Simulation at some level for X-D Monte Carlo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vector$simulate_level(niter0, iter1 = 1, level = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{niter0}}{Number of iterations at the lower level}

\item{\code{iter1}}{Iteration at the upper level}

\item{\code{level}}{Level of the calculation (ignored, it always returns a vector
to facilitate the mutates)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Vector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Vector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Vector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
